moved GenRegDescsAnno to utils/Annotations
