+++ src/main/scala/regmapper/Annotation.scala
@@ -2,7 +2,6 @@
 
 package freechips.rocketchip.regmapper
 
-import chisel3.core.annotate
 import chisel3.experimental.{ChiselAnnotation, RawModule, RunFirrtlTransform}
 import firrtl.annotations._
 import firrtl.{CircuitForm, CircuitState, LowForm, Transform}
@@ -36,26 +35,12 @@ case class RegistersSer(
   regFields: Seq[RegFieldSer]
 )
 
-/**
-  * Firrtl annotation
-  *
-  * @param target
-  * @param regMappingSer
-  */
 case class RegFieldDescMappingAnnotation(
   target: ModuleName,
   regMappingSer: RegistersSer) extends SingleTargetAnnotation[ModuleName] {
     def duplicate(n: ModuleName): RegFieldDescMappingAnnotation = this.copy(target = n)
 }
 
-/**
-  * Chisel Annotation
-  *
-  * This creates the firrtl annotation. The Chisel Annotation is converted to a FIRRTL annotation by the toFirrtl class
-  *
-  * @param target
-  * @param regMappingSer
-  */
 case class RegMappingChiselAnnotation(
   target: RawModule,
   regMappingSer: RegistersSer) extends ChiselAnnotation with RunFirrtlTransform {
@@ -74,103 +59,5 @@ class RegMappingDumpTransform extends Transform {
       case RegFieldDescMappingAnnotation(t, desc) => println(desc)
     }
     state
-  }
-}
-
-object GenRegDescsAnno {
-  // Each class which has regmaps has instance counters
-  private var instanceCounters = scala.collection.mutable.Map[String, Int]()
-
-  def addInstanceCounter(key: String) : Unit = {
-    instanceCounters +=  ( key -> 0)
-  }
-
-  def getInstanceCount(name: String, baseAddress: BigInt) : Int = {
-    val nameAddress = s"${name}.${baseAddress}"
-
-    if (! (instanceCounters isDefinedAt nameAddress)) {
-      addInstanceCounter(nameAddress)
-    }
-    val cnt = instanceCounters(nameAddress)
-    val newCnt = cnt + 1
-    instanceCounters(nameAddress) = newCnt
-    cnt
-  }
-
-  def makeRegMappingSer(rawModule: RawModule,
-    moduleName: String,
-    baseAddress: BigInt,
-    width: Int,
-    byteOffset: Int,
-    bitOffset: Int,
-    regField: RegField): RegFieldSer = {
-
-    val anonRegFieldName = s"unnamedRegField${byteOffset.toHexString}_${bitOffset}"
-    val regFieldName = regField.desc.map {_.name}.getOrElse("")
-    val selectedRegFieldName = if (regFieldName.isEmpty /* selectedName.isEmpty */) anonRegFieldName else regFieldName
-    val map = Map[BigInt, (String, String)]() // TODO
-
-    val byteOffsetHex = s"0x${byteOffset.toInt.toHexString}"
-
-    val desc = regField.desc
-
-    val regFieldDescSer = RegFieldDescSer(
-      byteOffset = byteOffsetHex,
-      bitOffset = bitOffset,
-      bitWidth = width,
-      name = selectedRegFieldName,
-      desc = desc.map {_.desc}.getOrElse("None"),
-      group = desc.map {_.group.getOrElse("None")}.getOrElse("None"),
-      groupDesc = desc.map {_.groupDesc.getOrElse("None")}.getOrElse("None"),
-      accessType = desc.map {_.access.toString}.getOrElse("None"),
-      wrType = desc.map(_.wrType.toString).getOrElse("None"),
-      rdAction = desc.map(_.rdAction.toString).getOrElse("None"),
-      volatile = desc.map(_.volatile).getOrElse(false),
-      hasReset = desc.map {_.reset != None }.getOrElse(false),
-      resetValue = desc.map{_.reset.getOrElse(BigInt(0))}.getOrElse(BigInt(0)),
-      enumerations = map
-    )
-
-    RegFieldSer(
-      moduleName, //selectedName,
-      regFieldDescSer
-     )
-  }
-
-
-  def anno(
-    rawModule: RawModule,
-    baseAddress: BigInt,
-    mapping: RegField.Map*): Seq[RegField.Map] = {
-
-    val moduleName = rawModule.name
-    val baseHex = s"0x${baseAddress.toInt.toHexString}"
-    val displayName = s"${moduleName}.${baseHex}"
-    val instanceCounter = GenRegDescsAnno.getInstanceCount(moduleName, baseAddress)
-
-    val regFieldSers = mapping.flatMap {
-      case (byteOffset, seq) =>
-        seq.map(_.width).scanLeft(0)(_ + _).zip(seq).map { case (bitOffset, regField) =>
-          makeRegMappingSer(
-            rawModule,
-            moduleName,
-            baseAddress,
-            regField.width,
-            byteOffset,
-            bitOffset,
-            regField
-          )
-        }
-    }
-
-    val registersSer = RegistersSer(
-      displayName = moduleName,
-      instanceCounter = instanceCounter,
-      baseAddress = baseAddress,
-      regFields = regFieldSers // Seq[RegFieldSer]()
-    )
-
-    annotate(RegMappingChiselAnnotation(rawModule, registersSer))
-    mapping
   }
 }
